+++ cmd/updater/main.go
@@ -1,7 +1,11 @@
+//nolint:mnd // quiet linter
+// SPDX-FileCopyrightText: Copyright © 2025 Ross Smith II <ross@smithii.com>
+// SPDX-License-Identifier: MIT
+
+//go:build !wasip1
+
 // package updater checks if there are any updates to snippetted code, and
 // provides diffs of any changes found.
-//
-//nolint:mnd // quiet linter
 package main
 
 import (