+++ dir_test.go
@@ -41,3 +41,30 @@ func TestReadDir(t *testing.T) {
 		t.Fatalf("ReadDir %s: golang directory not found", dirname)
 	}
 }
+
+func TestDirEntry(t *testing.T) {
+	dirname := "."
+	list, err := compat.ReadDir(dirname)
+	if err != nil {
+		t.Fatalf("ReadDir %s: %v", dirname, err)
+	}
+
+	for _, dir := range list {
+		info, err := dir.Info()
+		if err != nil {
+			t.Fatalf("ReadDir %s: %v", dirname, err)
+		}
+		if info == nil {
+			t.Fatalf("ReadDir %s: %v", dirname, "info is nil")
+		}
+		if info.Name() != dir.Name() {
+			t.Fatalf("ReadDir %s: Name(): got %v; want %v", dirname, info.Name(), dir.Name())
+		}
+		if info.IsDir() != dir.IsDir() {
+			t.Fatalf("ReadDir %s: IsDir(): got %v; want %v", dirname, info.IsDir(), dir.IsDir())
+		}
+		if info.Mode().Type() != dir.Type() {
+			t.Fatalf("ReadDir %s: Type(): got %v; want %v", dirname, info.Mode().Type(), dir.Type())
+		}
+	}
+}